1.15.1 scythe package refactor
