+++ src/main/java/com/lothrazar/cyclicmagic/module/ItemModule.java
@@ -50,7 +50,6 @@ import com.lothrazar.cyclicmagic.item.ItemPistonWand;
 import com.lothrazar.cyclicmagic.item.ItemPlayerLauncher;
 import com.lothrazar.cyclicmagic.item.ItemProspector;
 import com.lothrazar.cyclicmagic.item.ItemRotateBlock;
-import com.lothrazar.cyclicmagic.item.ItemScythe;
 import com.lothrazar.cyclicmagic.item.ItemSleepingMat;
 import com.lothrazar.cyclicmagic.item.ItemSoulstone;
 import com.lothrazar.cyclicmagic.item.ItemSpawnInspect;
@@ -132,6 +131,7 @@ import com.lothrazar.cyclicmagic.item.mobs.ItemHorseUpgrade;
 import com.lothrazar.cyclicmagic.item.mobs.ItemHorseUpgrade.HorseUpgradeType;
 import com.lothrazar.cyclicmagic.item.mobs.ItemVillagerMagic;
 import com.lothrazar.cyclicmagic.item.random.ItemRandomizer;
+import com.lothrazar.cyclicmagic.item.scythe.ItemScythe;
 import com.lothrazar.cyclicmagic.item.storagesack.ItemStorageBag;
 import com.lothrazar.cyclicmagic.item.tiletransporter.ItemChestSack;
 import com.lothrazar.cyclicmagic.item.tiletransporter.ItemChestSackEmpty;
